refactor - add repository setup class
